refactor: 자동차이름, 레이싱에 필요한 상수 각 파일로 분리
