+++ src/constant/carName.js
@@ -1,0 +1,4 @@
+export const CAR_NAME = {
+  LENGTH: 5,
+  MININUM: 0,
+};
