+++ src/constant/racing.js
@@ -1,8 +1,3 @@
-export const CAR_NAME = {
-  LENGTH: 5,
-  MININUM: 0,
-};
-
 export const RACING_COUNT = {
   GO_FORWARD: 4,
   MININUM: 0,
@@ -12,3 +7,5 @@ export const NUMBER = {
   MIN: 0,
   MAX: 9,
 };
+
+export const PROGRESS_SING = "-";
